% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_news_scraper.R
\name{mod_news_scraper_server}
\alias{mod_news_scraper_server}
\title{news_scraper Server Functions}
\usage{
mod_news_scraper_server(id, selected_product_r)
}
\description{
news_scraper Server Functions
}
