% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_product_selector.R
\name{mod_product_selector_server}
\alias{mod_product_selector_server}
\title{product_selector Server Functions}
\usage{
mod_product_selector_server(id, market)
}
\description{
product_selector Server Functions
}
